+++ AnimatableImage.js
